+++ src/plotting.jl
@@ -105,6 +105,8 @@ See also [`plot_basins_attractors_curves`](@ref) and
   plot of each basin fraction
 - `separatorwidth = 1, separatorcolor = "white"`: adds a line separating the fractions
   if the style is `:band`
+- `filler = NaN`: filler value to use for basin fractions for attractor IDs that do not exist at a
+  continuation step when the style is `:lines` (filler is always `0` for `:band` style).
 - `axislegend_kwargs = (position = :lt,)`: propagated to `axislegend` if a legend is added
 - `series_kwargs = NamedTuple()`: propagated to the band or scatterline plot
 - Also all [common plotting keywords](@ref common_plot_kwargs).
@@ -134,8 +136,9 @@ export plot_attractors_curves, plot_attractors_curves!
 Same as in [`plot_basins_curves`](@ref) but visualize any arbitrary quantity characterizing
 the continuation. Hence, the `continuation_info` is of exactly the same format as
 `fractions_cont`: a vector of dictionaries, each dictionary mapping attractor IDs to real numbers.
-`continuation_info` is meant to accompany `attractor_info` in [`plot_attractors_curves`](@ref).
-To produce `continuation_info` from `attractor_info` you can do something like:
+`continuation_info` is meant to accompany `attractors_cont` in [`plot_attractors_curves`](@ref).
+
+To produce `continuation_info` from `attractors_cont` you can do something like:
 
 ```julia
 continuation_info = map(attractors_cont) do attractors
@@ -146,7 +149,7 @@ with `f` your function of interest that returns a real number.
 
 ## Keyword arguments
 
-- `slines_kwargs = (linewidth = 1,)`: named tuple of arguments propagated
+- `series_kwargs`: named tuple of arguments propagated
   to `Makie.scatterlines!` that plots the curves.
 - Also all [common plotting keywords](@ref common_plot_kwargs).